requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' --strip-extras '.\requirements.in'
#
beautifulsoup4==4.14.3
    # via -r requirements.in
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
google-play-scraper==1.2.7
    # via -r requirements.in
greenlet==3.3.2
    # via sqlalchemy
idna==3.11
    # via requests
lxml==6.0.2
    # via -r requirements.in
numpy==2.4.2
    # via pandas
pandas==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.47
    # via -r requirements.in
tenacity==9.1.4
    # via -r requirements.in
tqdm==4.67.3
    # via -r requirements.in
truststore==0.10.4
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   sqlalchemy
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
